==> Human_Divide/03_img_minist_pca.ipynb <==
#### 【 이미지 기계학습 】

- 이미지 전처리
    * 입력 형태를 맞추는 전처리는 모두 동일 적용
    * 분포를 흔드는 랜덤 변환은 train만 적용
    * train/valid/test 모두 적용
        - 리사이징, 크롭, 배경 제거, 색공간 변경, 픽셀 스케일링 변경
    * train만 진행
        - 데이터 증강용 작업 : Random flip, Random rotation, Color jitter ㅡ Noise 추가  


## =================================================================
## [1-1] 모듈 로딩
## =================================================================
##- 기본 모듈
import numpy as np
import pandas as pd

##- ML 데이터셋 관련 
from sklearn.model_selection import train_test_split, StratifiedKFold, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA

##- ML 학습 알고리즘 관련
from sklearn.linear_model import LogisticRegression

##- ML 성능지표 관련
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

## =================================================================
## [1-2] 데이터 준비
## =================================================================
##- 데이터 설정
CSV_PATH = "./Data/img_face.csv"

##- 데이터 로딩 
mnistDF  = pd.read_csv(CSV_PATH) 

display(mnistDF.head())

[2] 피쳐와 타겟 & 학습용과 테스트용 데이터셋 준비<hr>

## =====================================================
## [2-1] 첫 컬럼: 라벨, 나머지 784개: 픽셀(28x28)
## =====================================================
y = mnistDF.iloc[:, -1]
X = mnistDF.iloc[:, :-1].astype(np.float32).values

print("X shape:", X.shape, "y shape:", y.shape)  # (N, 784), (N,)

## =====================================================
## [2-2] train/test 분리
## =====================================================
X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)


[3] 교차검증 통한 하이퍼파라미터 찾기<hr>

# =========================================================
# [3-1] Pipeline (누수 방지: PCA는 CV fold의 train에서만 fit)
# =========================================================
pipe = Pipeline(steps=[
    ("scaler", StandardScaler()),
    ("pca",    PCA(random_state=42)),
    ("clf",    LogisticRegression(max_iter=3000,solver="lbfgs" ))
])

# =========================================================
# [3-2] 교차검증 위한 학습 모델 하이퍼라라미터 
# =========================================================
param_grid = {
    # PCA 차원: 정수(주성분 개수) 또는 누적분산 비율(0~1)
    "pca__n_components": [0.90, 0.95, 0.97, 80, 120, 140, 160],
    # LR 규제 강도(C): 클수록 규제 약함
    "clf__C": [0.1, 1.0, 3.0, 10.0],
}

# =========================================================
# [3-3] 사용자 정의 교차검증 설정
# =========================================================
## 자동 CV는 내부에서 random_state 고정하지 않음 
## 재현성 위해 설정 
cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)


## ===================================================================================
## [3-4] 교차검증
## -> refit 매개변수 
##      * True  : 전체 train 데이터 + 최적 하이퍼 파라미터로 학습 후 ESM 반환
##      * False : X ===> 개발자가 직접 전체 train 데이터 + 최적 하이퍼 파라미터로 학습 잰행
## ===================================================================================
grid = GridSearchCV(
    estimator=pipe,
    param_grid=param_grid,
    scoring="accuracy",
    cv=cv,
    verbose=3,
    refit=True
)

grid.fit(X_train, y_train)

print("\n[BEST PARAMS]")
print(grid.best_params_)

print("[BEST CV SCORE]", grid.best_score_)


[2] PCA만 확인 <hr>

stdScaler = StandardScaler()
X_train_scaled = stdScaler.fit_transform(X_train)
X_test_scaled  = stdScaler.transform(X_test)

print(f'{X_train_scaled.shape} / {X_test_scaled.shape}')


pca = PCA(random_state=42,n_components=160)

X_train_pca = pca.fit_transform(X_train_scaled)
X_test_pca  = pca.transform(X_test_scaled)
print(f'pca_x_train : {X_train_pca.shape}')

print("Train PCA:", X_train_pca.shape)
print("Test  PCA :", X_test_pca.shape)
print("Explained variance:", pca.explained_variance_ratio_.sum())

import pandas as pd

train_pca_df = pd.DataFrame(
    X_train_pca,
    columns=[f"pc{i}" for i in range(X_train_pca.shape[1])]
)
train_pca_df["target"] = y_train.values

train_pca_df.to_csv("./Data/img_face_pca120_train.csv", index=False)

test_pca_df = pd.DataFrame(
    X_test_pca,
    columns=[f"pc{i}" for i in range(X_test_pca.shape[1])]
)
test_pca_df["target"] = y_test.values

test_pca_df.to_csv("./Data/img_face_pca120_test.csv", index=False)

import joblib

joblib.dump(stdScaler, "./Model/scaler.joblib")
joblib.dump(pca, "./Model/pca_160.joblib")
